+++ src/module/vue/components/sf-movesheetmoves.vue
@@ -118,11 +118,9 @@ CONFIG.IRONSWORN.emitter.on('highlightMove', async (targetMoveUuid) => {
 	clearSearch()
 	await nextTick()
 	const categoryWithMove = allCategories.value.find((moveCategory) =>
-		moveCategory.moveItems.has(targetMoveUuid ?? '')
+		moveCategory.moveUuids.includes(targetMoveUuid ?? '')
 	)
-	if (categoryWithMove) {
-		categoryWithMove.expandAndHighlightMove(targetMoveUuid)
-	}
+	categoryWithMove?.expandAndHighlightMove(targetMoveUuid)
 })
 </script>